chore: add warning when secret file exists (#1165)
(cherry picked from commit ea4dc45fe31787f9f49e549bd2f3576e450d2a6e)

+++ bin/create-secrets.sh
@@ -102,6 +102,14 @@ ironic_rabbitmq_password=$(generate_password 32)
 
 OUTPUT_FILE="/etc/genestack/kubesecrets.yaml"
 
+if [[ -f ${OUTPUT_FILE} ]]; then
+    echo "Error: ${OUTPUT_FILE} already exists. Please remove it before running this script."
+    echo "       This will replace an existing file and will lead to mass rotation, which is"
+    echo "       likely not what you want to do. If you really want to break your system, please"
+    echo "       make sure you know what you're doing."
+    exit 99
+fi
+
 cat <<EOF > $OUTPUT_FILE
 ---
 apiVersion: v1
